Preserve idempotency of docker::image loaded from tar
The docker::image resource looks for an existing image name in order to
determine whether to execute. The Harbor release tar includes all
Harbor images in one tar. A single image name is selected for use as the
docker::image title for offline installs. This name should exists in all
supported Harbor releases to preserve the dopcker::image resource
idempotency.